Move test cleanup hack to conftest.py
