+++ src/planingfsi/cli.py
@@ -41,8 +41,6 @@ def cli() -> None:
 )
 def run_planingfsi(post_mode: bool, plot_save: bool, plot_show: bool, new_case: bool) -> None:
     """Run the planingFSI solver."""
-    _cleanup_globals()
-
     simulation = Simulation.from_input_files("configDict")
 
     simulation.config.plotting.save = plot_save
@@ -70,8 +68,6 @@ def generate_mesh(
     mesh_dict: Optional[str], plot_show: bool, plot_save: bool, verbose: bool
 ) -> None:
     """Generate the initial mesh."""
-    _cleanup_globals()
-
     config = Config.from_file("configDict")
 
     if mesh_dict is not None:
@@ -86,29 +82,3 @@ def generate_mesh(
     mesh.display(disp=verbose)
     mesh.plot(show=plot_show, save=plot_save)
     mesh.write()
-
-
-def _cleanup_globals() -> None:
-    """Clear out class-global lists, which causes trouble when running the program multiple times
-    during the same testing session.
-
-    Todo:
-        * This should be removed after the globals are factored out.
-
-    """
-    from planingfsi.fe.felib import Element
-    from planingfsi.fe.felib import Node
-    from planingfsi.fe.femesh import Curve
-    from planingfsi.fe.femesh import Line
-    from planingfsi.fe.femesh import Point
-    from planingfsi.fe.femesh import Shape
-    from planingfsi.fe.substructure import Substructure, FlexibleSubstructure
-
-    Shape._Shape__all = []  # type: ignore
-    Point._Point__all = []  # type: ignore
-    Curve._Curve__all = []  # type: ignore
-    Line._Line__all = []  # type: ignore
-    Node._Node__all = []  # type: ignore
-    Element._Element__all = []  # type: ignore
-    Substructure._Substructure__all = []  # type: ignore
-    FlexibleSubstructure._FlexibleSubstructure__all = []  # type: ignore
